Platform/DeveloperBox: omit TPM from DT when building without TPM support
The recently added support for TPM2 measured boot added a description of
the TPM to the device tree, but failed to take the build configuration
into account, and so it adds it unconditionally.

Fix this, by #define'ing a TPM2_ENABLE CPP macro that can be referenced
in the device tree source file.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>