KabylakeSiliconPkg/DxeAslUpdateLib: Cleans up APIs
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=2536

Callers of DxeAslUpdateLib don't have to call InitializeAslUpdateLib()
but the library itself runs it internally. This commit deletes
the redundant calls in caller side and makes it an internal call.
LocateAcpiTableByOemTableId() is unreferenced externally
so this commit makes it an internal call.
PSS_PACKAGE_LAYOUT and AML_RESRC_TEMP_END_TAG are both unreferenced
so they are removed.
AslUpdateLib.h is cleaned up accordingly.

Signed-off-by: Miki Shindo <miki.shindo@intel.com>
Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Prince Agyeman <prince.agyeman@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Acked-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Ray Ni <ray.ni@intel.com>
Reviewed-by: Sai Chaganty <rangasai.v.chaganty@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>