Platform/RaspberryPi: Documentation updates
This cleans up the (related) Readme.md files for both platforms.

- The RPi3 target also supports the 3A+, so let's document it.
(The 3A+ lacks the ethernet port and has less RAM)

- Clarify RPi3 is for BCM2837-based variants, while RPi4 is
for the BCM2711-based variants.

- v1.2 (not older) actually uses the same BCM2837 SoC as
Pi 3, and is thus a 64-bit device that this firmware
easily supports as a minor variant of the Pi 3.

The difference from Pi 3 is choice of UART (PL011), no WiFi,
no BT and slower clock due to thermal problems caused by
older board design.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@arm.com>