Silicon/Synopsys/DesignWare: import eMMC DXE driver from EmbeddedPkg
Incorporate the driver for the DesignWare eMMC host controller that is
based on the obsolete MMC host controller protocol that is defined in
EmbeddedPkg.

This driver does not follow the UEFI driver model, and is only kept
around for its only users, which is the HiKey platform, which is
rapidly reaching obsolescence itself, at which point this driver may
be removed again.

To prevent inadvertent use in new platforms, add a PCD that needs to
be changed from its default value in order for the driver to be
functional.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>
Reviewed-by: Philippe Mathieu-Daude <philmd@redhat.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>