Silicon/NXP: Add I2c lib
Add an I2c library to control the i2c controllers to communicate
with I2c device.

We need this functionality in a lib because this is going to be used
in PrePeiCore sec module to get the System clock information from
devices connected to i2c (like fpga or clock generator)

In subsequent patches I2c functionality in I2cDxe driver would be removed
and I2cLib APIs would be used in I2cDxe.

The I2cLib differs from I2c Controller functionality in I2c Dxe in
several aspects:
- I2cLib doesn't assume that each I2c Transaction would be of minimum
  two operations
- I2cLib generates repeat start between operations which complies with
  PI I2c specs.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>