Platform/ARM/SgiPkg: Create platform specific dsc files
There are currently four platforms that are supported under SgiPkg and
all these four platforms have a common platform dsc file to allow these
to be built into a single binary. But the newer platforms that are
intended to be added to the SgiPkg differ to some extent on the software
programming interface. So in order to accommodate the newer platforms
into SgiPkg, there is shift from building single binary for all
platforms to independent binary for each platform.

As a first step towards this, create .dsc file for each supported
platform. The exiting .dsc file is converted to be a include file
that the platform specific .dsc files can include.

Signed-off-by: Aditya Angadi <aditya.angadi@arm.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@arm.com>