Silicon/Broadcom/BcmGenetDxe: Fix reliability issues
TFTP and HTTP downloads were failing. The problem was that the incorrect
PROD index was being written TX ring. The PROD index should be the
TxProdIndex (in the range [0-0xffff]), not the descriptor index (in the
range [0-num descs)).

Also, mod 0xffff is not the same as & 0xffff, so fix that as well.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>
Tested-by: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>