Platform/96Boards/96BoardsI2cDxe: connect I2C controllers at EndOfDxe
The 96boards I2C driver currently relies on the platform to connect
all controllers, or I2C peripherals will not be exposed if they are
not the active boot target. Since I2C peripherals are not boot targets
in the first place, but are used to expose things like random number
generators, let's connect the I2C controllers specifically at EndOfDxe
so that the devices living on it will be available regardless of the
boot policy.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>