Vlv2TbltDevicePkg/FmpDeviceLib: Add LastAttemptStatus compatibility
Makes the changes necessary for these library instances of
FmpDeviceLib to be compatible with new functions added recently
to FmpDeviceLib.

Two new functions were introduced in FmpDeviceLib to allow a
library instance to return a Last Attempt Status code during
check image and set image operations:
  1. FmpDeviceCheckImageWithStatus ( )
  2. FmpDeviceSetImageWithStatus ( )

FmpDxe (in FmpDevicePkg) will begin calling these new functions
instead of the previous functions. Therefore, this change:
  1. Adds these functions to Vlv2TbltDevicePkg implementations
  2. Moves the main functionality to these new functions
  3. Updates the old functions to call the new functions
     (for backward compatibility)

Note: As of this commit, the Vlv2TbltDevicePkg build is broken
due to:
  1. A required RngLib library instance not defined by the platform
  2. Other FMP libraries not being defined by the platform
     (e.g. FmpDependencyLib, FmpDependencyCheckLib, etc.)

Those changes were fixed locally to test the changes in this commit
but maintainers should make the proper changes for those issues.

Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Yi Qian <yi.qian@intel.com>
Cc: Zailiang Sun <zailiang.sun@intel.com>
Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>
Reviewed-by: Michael D Kinney <michael.d.kinney@intel.com>
Tested-by: Michael D Kinney <Michael.d.kinney@intel.com>